+++ Source/Lib/RAWCooked/RAWCooked.h
@@ -11,9 +11,10 @@
 
 //---------------------------------------------------------------------------
 #include <cstdint>
+#include <cstddef>
 //---------------------------------------------------------------------------
 
-typedef void(*write_file_call)(uint8_t*, uint64_t, void*);
+typedef void(*write_file_call)(uint8_t*, size_t, void*);
 
 class rawcooked
 {
